+++ .lycheeignore
@@ -1,3 +1,3 @@
 file:///*
 http://localhost:*
-https://arango.qubitpi.org/*
+https://arango.paion-data.dev/*
