+++ README.md
@@ -24,7 +24,7 @@ For this reason, Entity Webservice is suitable for a microservice architecture.
 Documentation
 -------------
 
-More information about Entity-WS can be found at [entity-ws.qubitpi.org](https://entity-ws.qubitpi.org/).
+More information about Entity-WS can be found at [entity-ws.paion-data.dev](https://entity-ws.paion-data.dev/).
 
 License
 -------
@@ -36,7 +36,7 @@ The use and distribution terms for [entity-ws]() are covered by the [Apache Lice
 [API Doc Badge]: https://img.shields.io/badge/Open%20API-Swagger-85EA2D.svg?style=for-the-badge&logo=openapiinitiative&logoColor=white&labelColor=6BA539
 [API Doc URL]: https://springdoc.org/
 
-[Docker Pulls Badge]: https://img.shields.io/docker/pulls/jack20191124/entity-ws?style=for-the-badge&logo=docker&color=2596EC
-[Docker Hub URL]: https://hub.docker.com/r/jack20191124/entity-ws
+[Docker Pulls Badge]: https://img.shields.io/docker/pulls/paiondatahub/entity-ws?style=for-the-badge&logo=docker&color=2596EC
+[Docker Hub URL]: https://hub.docker.com/r/paiondatahub/entity-ws
 
 [Java Version Badge]: https://img.shields.io/badge/Java-17-brightgreen?style=for-the-badge&logo=OpenJDK&logoColor=white
