+++ docker-compose.yml
@@ -1,4 +1,4 @@
-# Copyright 2025 Jiaqi Liu. All rights reserved.
+# Copyright 2025 Paion Data. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
 version: "3.9"
 services:
   entity-ws:
-    image: jack20191124/entity-ws
+    image: paiondatahub/entity-ws
     ports:
       - "8080:8080"
     environment:
