+++ docs/docs/development.mdx
@@ -3,7 +3,7 @@ sidebar_position: 2
 title: Development
 ---
 
-[//]: # (Copyright 2025 Jiaqi Liu. All rights reserved.)
+[//]: # (Copyright 2025 Paion Data. All rights reserved.)
 
 [//]: # (Licensed under the Apache License, Version 2.0 &#40;the "License"&#41;;)
 [//]: # (you may not use this file except in compliance with the License.)
@@ -54,12 +54,12 @@ The ArangoDB Web console should be accessible at http://localhost:8529 with the 
 :::tip
 
 For more documentation on navigating the Arango Web UI, please refer to
-[ArangoDB documentation](https://arango.qubitpi.org/stable/components/web-interface/)
+[ArangoDB documentation](https://arango.paion-data.dev/stable/components/web-interface/)
 
 :::
 
 ```console
-git clone git@github.com:QubitPi/entity-ws.git
+git clone git@github.com:paion-data/entity-ws.git
 cd entity-ws
 mvn clean package
