+++ pom.xml
@@ -10,12 +10,12 @@
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
 
-    <groupId>org.qubitpi</groupId>
+    <groupId>com.paiondata</groupId>
     <artifactId>entity-ws</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <name>entity-ws</name>
     <description>Spring Boot Starter based CRUD webservice</description>
-    <url>https://github.com/QubitPi/entity-ws</url>
+    <url>https://github.com/paion-data/entity-ws</url>
 
     <licenses>
         <license>
@@ -27,8 +27,8 @@
 
     <developers>
         <developer>
-            <name>Jiaqi Liu</name>
-            <url>https://github.com/QubitPi</url>
+            <name>Paion Data</name>
+            <url>https://github.com/paion-data</url>
         </developer>
     </developers>
