+++ src/main/java/com/paiondata/ws/entity/SwaggerConfig.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 Jiaqi Liu. All rights reserved.
+ * Copyright 2025 Paion Data. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.qubitpi.ws.entity;
+package com.paiondata.ws.entity;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -44,9 +44,9 @@ public class SwaggerConfig {
                                 .description("Entity Webservice REST API Interface")
                                 .contact(
                                         new Contact()
-                                                .name("Jiaqi Liu")
-                                                .url("https://qubitpi.org")
-                                                .email("jack20220723@gmail.com")
+                                                .name("Paion Data")
+                                                .url("https://github.com/paion-data")
+                                                .email("jiaqiliu@paion-data.dev")
                                 )
                                 .license(
                                         new License()
