+++ src/main/java/com/paiondata/ws/entity/arango/AbstractRestService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 Jiaqi Liu. All rights reserved.
+ * Copyright 2025 Paion Data. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.qubitpi.ws.entity.arango;
+package com.paiondata.ws.entity.arango;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -36,7 +36,7 @@ import java.util.Map;
 /**
  * {@link AbstractRestService} abstracts away the ArangoDB REST API request implementation.
  * <p>
- * See https://arango.qubitpi.org/stable/develop/http-api/ for more information.
+ * See https://arango.paion-data.dev/stable/develop/http-api/ for more information.
  */
 abstract class AbstractRestService {
 
@@ -168,7 +168,7 @@ abstract class AbstractRestService {
     /**
      * Sends an authentication request to Arango auth endpoint and returns a JWT token.
      * <p>
-     * See https://arango.qubitpi.org/stable/develop/http-api/authentication/ for more info
+     * See https://arango.paion-data.dev/stable/develop/http-api/authentication/ for more info
      *
      * @return a string representing the JWT token
      */
