+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-# Copyright 2025 Jiaqi Liu. All rights reserved.
+# Copyright 2025 Paion Data. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
