+++ src/test/java/com/paiondata/ws/entity/SwaggerIT.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 Jiaqi Liu. All rights reserved.
+ * Copyright 2025 Paion Data. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.qubitpi.ws.entity;
+package com.paiondata.ws.entity;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -22,12 +22,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.http.HttpStatus;
 
 /**
- * The entity webservice integration tests.
+ * Open API documentation integration tests.
  */
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class ApplicationIT {
+class SwaggerIT {
 
     @LocalServerPort
     private int port;
@@ -36,18 +37,11 @@ class ApplicationIT {
     private TestRestTemplate restTemplate;
 
     /**
-     * Make sure the context is creating controller.
+     * Make sure Swagger UI loads.
      */
     @Test
-    void contextLoads() {
-    }
-
-    /**
-     * Make sure actuator is working properly.
-     */
-    @Test
-    void testHealthcheck() {
-        assertThat(this.restTemplate.getForObject("http://localhost:" + this.port + "/actuator/health", String.class))
-                .isEqualTo("{\"status\":\"UP\"}");
+    void uiTest() {
+        assertThat(this.restTemplate.getForEntity("http://localhost:" + port + "/swagger-ui/index.html",
+                String.class).getStatusCode()).isEqualTo(HttpStatus.OK);
     }
 }
